python/requirements.txt
fastapi
uvicorn
python-dotenv
runpod
torch
transformers
wandb
tqdm
numpy
scikit-learn
pandas